底层重构
删除了错误理解的dto层，移动至service层，将大部分new操作，改为使用aotowired的单例模式，修改金额为double，添加图片地址的长度判断，减少服务器报错。

+++ src/main/java/com/web/entity/Content.java
@@ -5,7 +5,7 @@ package com.web.entity;
  */
 public class Content {
     private int id;
-    private int price;
+    private Double price;
     private String title;
     private String pic;
     private String summary;
@@ -28,11 +28,11 @@ public class Content {
     }
 
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
